requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.115.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.0
groq==1.1.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==1.51.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.3.2
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.5
sniffio==1.3.1
starlette==0.38.6
tenacity==9.1.4
tqdm==4.67.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
